+++ samza-test/src/test/java/org/apache/samza/test/table/TestRemoteTableEndToEnd.java
@@ -54,8 +54,7 @@ import org.apache.samza.serializers.NoOpSerde;
 import org.apache.samza.table.Table;
 import org.apache.samza.table.descriptors.CachingTableDescriptor;
 import org.apache.samza.table.descriptors.GuavaCacheTableDescriptor;
-import org.apache.samza.table.remote.RemoteReadWriteTable;
-import org.apache.samza.table.remote.RemoteReadableTable;
+import org.apache.samza.table.remote.RemoteTable;
 import org.apache.samza.table.descriptors.RemoteTableDescriptor;
 import org.apache.samza.table.remote.TableRateLimiter;
 import org.apache.samza.table.remote.TableReadFunction;
@@ -80,7 +79,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.withSettings;
 
 
-public class TestRemoteTable extends AbstractIntegrationTestHarness {
+public class TestRemoteTableEndToEnd extends AbstractIntegrationTestHarness {
 
   static Map<String, List<EnrichedPageView>> writtenRecords = new HashMap<>();
 
@@ -184,7 +183,7 @@ public class TestRemoteTable extends AbstractIntegrationTestHarness {
     String profiles = Base64Serializer.serialize(generateProfiles(count));
 
     int partitionCount = 4;
-    Map<String, String> configs = TestLocalTable.getBaseJobConfig(bootstrapUrl(), zkConnect());
+    Map<String, String> configs = TestLocalTableEndToEnd.getBaseJobConfig(bootstrapUrl(), zkConnect());
 
     configs.put("streams.PageView.samza.system", "test");
     configs.put("streams.PageView.source", Base64Serializer.serialize(pageViews));
@@ -266,8 +265,8 @@ public class TestRemoteTable extends AbstractIntegrationTestHarness {
     future.completeExceptionally(new RuntimeException("Expected test exception"));
     doReturn(future).when(reader).getAsync(anyString());
     TableRateLimiter rateLimitHelper = mock(TableRateLimiter.class);
-    RemoteReadableTable<String, ?> table = new RemoteReadableTable<>(
-        "table1", reader, rateLimitHelper, Executors.newSingleThreadExecutor(), null);
+    RemoteTable<String, String> table = new RemoteTable<>("table1", reader, null,
+        rateLimitHelper, null, Executors.newSingleThreadExecutor(), null);
     table.init(createMockContext());
     table.get("abc");
   }
@@ -280,9 +279,32 @@ public class TestRemoteTable extends AbstractIntegrationTestHarness {
     future.completeExceptionally(new RuntimeException("Expected test exception"));
     doReturn(future).when(writer).putAsync(anyString(), any());
     TableRateLimiter rateLimitHelper = mock(TableRateLimiter.class);
-    RemoteReadWriteTable<String, String> table = new RemoteReadWriteTable<String, String>(
+    RemoteTable<String, String> table = new RemoteTable<String, String>(
         "table1", reader, writer, rateLimitHelper, rateLimitHelper, Executors.newSingleThreadExecutor(), null);
     table.init(createMockContext());
     table.put("abc", "efg");
+  }
+
+  @Test
+  public void testUninitializedWriter() {
+    TableReadFunction<String, String> reader = mock(TableReadFunction.class);
+    TableRateLimiter rateLimitHelper = mock(TableRateLimiter.class);
+    RemoteTable<String, String> table = new RemoteTable<String, String>(
+        "table1", reader, null, rateLimitHelper, null, Executors.newSingleThreadExecutor(), null);
+    table.init(createMockContext());
+    int failureCount = 0;
+    try {
+      table.put("abc", "efg");
+    } catch (SamzaException ex) {
+      ++failureCount;
+    }
+    try {
+      table.delete("abc");
+    } catch (SamzaException ex) {
+      ++failureCount;
+    }
+    table.flush();
+    table.close();
+    Assert.assertEquals(2, failureCount);
   }
 }
